pyproject.toml
[tool.poetry]
name = "pfa_columnar_analyzer"
version = "1.0.0"
description = ""
authors = ["Francesco Ivone <francesco.ivone@cern.ch>"]
packages = [
    { include = "*.py", from = "./tools"}
]    

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
awkward = "2.0.0"
matplotlib = "3.5.2"
numpy = "1.22.1"
pandas = "1.3.5"
seaborn = "0.11.2"
uproot = "5.0.0"
guppy3 = "^3.1.2"
numba = "^0.56.4"
icc-rt = "^2020.0.133"
pyyaml = "^6.0"
scipy = "^1.10.1"
tqdm = "^4.65.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
